requirements.txt
Flask
isort
black
pytest
psycopg2-binary
fastapi